com: introduce take_stream on PrinterCon trait
The previous idea was to have PrinterCon (the main connection struct)
serve as the Reader and Writer to the underlying socket/stream/… This
turned out to be impractical down the line, as it disallowed splitting
of streams and keeping ownership of the connection struct (with all
their helpers or additional information).

With take_stream from PrinterCon it is now possible to split of
communication with the device from the main device functionality while
keeping basic access to the underlying connection.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>